Add delete criteria

+++ create_test.go
@@ -100,8 +100,20 @@ func TestCreateStringArray(t *testing.T) {
 		if !assert.NoError(t, err, "insert fail using "+target.layer) {
 			return
 		}
+		count++
 		log.Log.Debugf("Delete of records done")
-		assert.Equal(t, int64(101), dr)
+		tId := "TEST" + strconv.Itoa(count)
+		list = append(list, []any{tId, "Tom", "Terminal"})
+		err = id.Insert(testCreationTable, &def.Entries{Fields: []string{"Id", "Name", "FirstName"},
+			Values: list})
+		if !assert.NoError(t, err, "insert fail using "+target.layer) {
+			return
+		}
+		dr, err = id.Delete(testCreationTable, &def.Entries{Criteria: "Id=" + tId})
+		if !assert.NoError(t, err, "delete fail using "+target.layer) {
+			return
+		}
+		assert.Equal(t, int64(1), dr)
 		if target.layer != "adabas" {
 			deleteTable(t, id, testCreationTable, target.layer)
 		}
